+++ Sources/BaseFoundationiOS/Common/Extensions/ArrayExtension.swift
@@ -30,3 +30,10 @@ extension Array where Element: Copying {
         return copiedArray
     }
 }
+
+extension Sequence where Element: Hashable {
+    public func uniqued() -> [Element] {
+        var set = Set<Element>()
+        return filter { set.insert($0).inserted }
+    }
+}